requirements.txt
Flask==3.0.0
Flask-Session==0.5.0
Flask-Limiter==3.5.0
pyotp==2.9.0
qrcode==7.4.2
Pillow==10.1.0
python-dotenv==1.0.0
cryptography==41.0.7
redis==5.0.1
APScheduler==3.10.4
Werkzeug==3.0.1
